refactor: extract CircuitBreaker from capture to internal/circuit
Phase 3 of capture god package split. Moves CircuitBreaker struct,
HealthResponse, RateLimitResponse types, and all constants to
internal/circuit/. Capture retains thin delegation methods and
type aliases for zero external API changes.

capture: 44 → 36 non-test files, ~6,600 → ~4,200 LOC (37% reduction).
