+++ internal/capture/rate_limit.go
@@ -1,9 +1,6 @@
-// Purpose: Owns rate_limit.go runtime behavior and integration logic.
-// Docs: docs/features/feature/backend-log-streaming/index.md
-
 // rate_limit.go — Capture delegation methods for rate limiting and circuit breaker.
-// Delegates to CircuitBreaker sub-struct. These methods preserve the existing
-// Capture API so callers (helpers.go, handlers, tests) don't need to change.
+// Delegates to CircuitBreaker sub-struct (aliased from internal/circuit).
+// These methods preserve the existing Capture API so callers don't need to change.
 package capture
 
 import (
@@ -12,24 +9,6 @@ import (
 
 	"github.com/dev-console/dev-console/internal/util"
 )
-
-// HealthResponse is returned by GET /health endpoint.
-type HealthResponse struct {
-	CircuitOpen bool   `json:"circuit_open"`
-	OpenedAt    string `json:"opened_at,omitempty"`
-	CurrentRate int    `json:"current_rate"`
-	Reason      string `json:"reason,omitempty"`
-}
-
-// RateLimitResponse is the 429 response body.
-type RateLimitResponse struct {
-	Error        string `json:"error"`
-	Message      string `json:"message"`
-	RetryAfterMs int    `json:"retry_after_ms"`
-	CircuitOpen  bool   `json:"circuit_open"`
-	CurrentRate  int    `json:"current_rate"`
-	Threshold    int    `json:"threshold"`
-}
 
 // RecordEvents delegates to CircuitBreaker.
 func (c *Capture) RecordEvents(count int) {
